Added enemy weapon and fire parameters to EnemyDetailsSO.cs. Created EnemyWeaponAI script. Configured the layers that the enemy can see and enemy bullets can interact with using a LayerMask. This class checks whether the player is in line of sight of the enemy and if the enemys ammo has range to reach the player.
Configured the Enemy.prefab to add the EnemyWeaponAI and related firing and reloading components. Modified the Enemy class to account for the changes, and set the enemy starting weapon on enemy initialization. Fixed AnimateEnemy.cs by removing the temporary code to set enemy aim direction and connected it to the enemys AimWeaponEvent to allow enemy to turn towards the player and not only left and right.