Created needed struct for enemy health and added necessary information to EnemyDetailsSO.cs. Configured health details for all enemies. Created the HealthEvent that will be invoked on health changes and added it to the enemy prefab. Made necessary changes to Health.cs and connected it with the HealthEvent. Made necessary changes in the Ammo.cs to make ammunition deal damage.
Created the DestroyedEvent that will make enemies and objects that reach zero health destroyed, and the Destroyed component. Added them to the Enemy prefab and did necessary changes to Enemy.cs so that enemies can take damage and be destroyed when their health reaches zero.

Made changes to InstantiatedRoom.cs so that the rooms can be unlocked and room collider can be re-enabled after the enemies in the room are cleared.

Created a new event in the StaticEventHandler.cs to help keep track of enemies in a given room.

Made changes to EnemySpawner.cs so that enemies are subscribed to the DestroyedEvent on creation. On an enemy destruction, the current enemy count is reduced and it is checked whether the current enemy count is zero and whether all designated enemies have spawned. If so, the game state is updated accordingly, and the doors are unlocked.

Player can now damage the and destroy the enemies, and progress through the dungeon rooms! The enemies still cannot deal damage to the player, and it will be implemented next.